refactor(architecture): share add contact components
