--- features/flow-contacts-add-contact/src/index.ts
@@ -1,4 +1,0 @@
-export * from "./screens/AddContact";
-export * from "./screens/AddContact/web";
-export { ContactNameInput } from "./screens/AddContact/components/ContactNameInput/ContactNameInput.web";
-export { ContactsAddContactNamingDisclaimer } from "./screens/AddContact/ContactsAddContactNamingDisclaimer.web";
