+++ features/flow/flow-contacts-add-contact/src/components/ContactNameInput/ContactNameInput.native.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Box, Text, TextInput } from "@ledgerhq/lumen-ui-rnative";
 import { DeleteCircleFill } from "@ledgerhq/lumen-ui-rnative/symbols";
-import { CONTACT_NAME_MAX_LENGTH } from "../../model/constants";
+import { CONTACT_NAME_MAX_LENGTH } from "./constants";
 
 type ContactNameInputProps = Readonly<{
   value: string;
