+++ features/flow/flow-contacts-add-contact/src/index.ts
@@ -1,0 +1,4 @@
+export * from "./screens/AddContact";
+export * from "./screens/AddContact/web";
+export { ContactNameInput } from "./components/ContactNameInput/ContactNameInput.web";
+export { ContactsAddContactNamingDisclaimer } from "./components/ContactsAddContactNamingDisclaimer.web";
