+++ features/flow/flow-contacts-add-contact/src/screens/AddContact/index.ts
@@ -1,5 +1,5 @@
 export type { ContactCreationInput, ContactCreationPort } from "./model/ports";
-export { CONTACT_NAME_MAX_LENGTH } from "./model/constants";
+export { CONTACT_NAME_MAX_LENGTH } from "../../components/ContactNameInput/constants";
 export type { AddContactViewModel } from "./model/types";
 export { createAddContactViewModel } from "./model/viewModel";
 export { createAddContactController, type AddContactController } from "./model/controller";
