+++ pnpm-workspace.yaml
@@ -2,7 +2,6 @@ packages:
   - "apps/*"
   - "apps/wallet-cli/npm/*"
   - "e2e/*"
-  - "features/*"
   - "features/flow/*"
   - "features/platform/*"
   - "libs/*"
